requirements.txt
langgraph==0.1.5
typer==0.12.0
pytest==8.3.2
docker==7.0.3
selenium==4.21.0
playwright==1.46.0
faiss-cpu==1.8.0
numpy==1.26.4
json==0.9.6
pathlib==1.0.1
